+++ src/main/java/emis.command/PrintCommand.java
@@ -1,0 +1,17 @@
+package main.java.emis.command;
+import main.java.emis.*;
+
+public class PrintCommand extends Command {
+    public PrintCommand() {
+    }
+
+    @Override
+    public void execute(TaskList t, Ui ui, Storage s) {
+        t.printList();
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
