+++ src/main/java/emis.exceptions/EmisException.java
@@ -1,0 +1,7 @@
+package main.java.emis.exceptions;
+
+public class EmisException extends Exception {
+    public EmisException(String str) {
+        super(str);
+    }
+}
